refactor: unexport NewBoolean and fix BooleanToBool pointer equality

+++ machine/vm_state.go
@@ -62,15 +62,67 @@ import (
 //	│ envPooled    │ ✓              │ ✗ (=false)  │ ✓ or false(shared)  │ ✓                │
 //	└──────────────┴────────────────┴─────────────┴─────────────────────┴──────────────────┘
 type vmState struct {
-	env          *environment.EnvironmentFrame
-	template     *NativeTemplate
-	singleValue  values.Value   // value register: single value (fast path, no allocation)
-	multiValues  MultipleValues // value register: multiple values (only for R7RS values/call-with-values)
+	env      *environment.EnvironmentFrame
+	template *NativeTemplate
+	// singleValue and multiValues form a split value register.
+	//
+	// Only one field is live at a time — the other is always nil.
+	// This is a performance optimisation: storing a single value in singleValue
+	// avoids the []Value{v} allocation that SetValues would otherwise cause on
+	// every load, store, and arithmetic instruction.
+	//
+	// INVARIANT: the two fields are mutually exclusive.
+	//   SetValue  sets singleValue and nils multiValues.
+	//   SetValues sets multiValues and nils singleValue (except for the
+	//   len==1 fast path, which delegates to SetValue).
+	//
+	// multiValues is only populated by R7RS (values ...) and is consumed by
+	// call-with-values. All other instructions operate on singleValue.
+	singleValue  values.Value   // value register: single value (fast path)
+	multiValues  MultipleValues // value register: multiple values (R7RS values/call-with-values only)
 	evals        *Stack         // evaluation stack, holds intermediate values during execution
 	pc           int
 	windingStack WindingStack // R7RS dynamic-wind extent tracking
 	promptTag    *PromptTag   // prompt tag for continuation prompts
-	threadID     uint64       // SRFI-18 thread identity: 0 = primordial thread
+	// threadID is the SRFI-18 thread identity (0 = primordial thread).
+	//
+	// Thread identity is split across two representations:
+	//
+	//   vmState.threadID (uint64)       — propagates into continuations
+	//   MachineContext.thread (*Thread)  — stays on the live context only
+	//
+	// Why two fields:
+	//
+	//   threadID lives in vmState, which is embedded by both MachineContext and
+	//   MachineContinuation. Every SaveContinuation stamps the current threadID
+	//   into the saved frame. Continuations need it for cross-thread rejection:
+	//
+	//     - applyComposableContinuation (machine_context.go) compares
+	//       mc.threadID against cc.threadID.
+	//     - call/cc escape closures (prim_control.go, prim_exit.go) compare
+	//       the invoking context's ThreadID() against the capturing thread's.
+	//     - ComposableContinuation stores its own threadID copy, also set from
+	//       mc.ThreadID() at capture time (prim_prompt.go, prim_control.go).
+	//
+	//   thread (*values.Thread) is NOT in vmState because continuations never
+	//   need the Scheme object — only the numeric ID for comparison. Storing a
+	//   pointer in vmState would add heap pressure to every continuation frame
+	//   for no benefit. The Scheme object is needed only by:
+	//
+	//     - current-thread (prim_threads.go:83) — returns mc.Thread() to Scheme.
+	//     - thread-specific mutation (prim_threads.go:405) — needs the owner.
+	//     - eval sub-contexts (prim_eval.go) — propagates thread via SetThread().
+	//
+	// Invariant: when MachineContext.thread is non-nil,
+	//   thread.ID() == threadID.
+	// Enforced by SetThread(), which is the single write path for both fields.
+	//
+	// Propagation:
+	//   NewSubContext      — copies both threadID and thread from parent.
+	//   NewThreadSubContext — sets both via SetThread(newThread).
+	//   SaveContinuation   — stamps threadID into the continuation frame.
+	//   Restore            — does NOT restore threadID (invoking thread keeps its own).
+	threadID uint64
 	// callDepth caches the continuation chain length to avoid O(d) traversals.
 	//
 	// On MachineContext: number of frames in the cont chain (mc.cont → ... → nil).
@@ -81,10 +133,9 @@ type vmState struct {
 	//   Set once at creation time, never mutated. A root frame (parent == nil)
 	//   has callDepth 0; each level adds 1.
 	//
-	// WARNING: callDepth is uint64. Subtracting from it can silently wrap to
-	// 2^64-1. All depth computation must use the parent pointer (which is nil-safe)
+	// All depth computation should use the parent pointer (which is nil-safe)
 	// rather than arithmetic on callDepth. See NewMachineContinuationFromMachineContext.
-	callDepth uint64
+	callDepth int
 	// envPooled is true when env was acquired from the envFramePool (Apply
 	// copy path). When the env is about to be overwritten (RestoreAndRelease,
 	// Restore), this flag tells us it's safe to return it to the pool.
